[thenakliman] #6 #8 Added table for user and roles and db framework
Following tasks has been done in this patch
    1. Update table "users" and "authority"(role).
    2. Added framework for the database connection.

Flyway has been used for database migrations and an initial
structure for the tables has been modified. Later on as per
requirement this structure could be changed.

For database ORM JPA has been used. There are no test cases for
the JPA, an in-memory database could be used for unit testing
purpose.

There are a few workaround(due to limited knowledge) in the test
cases, NOTE has been put, those workaround has to be fixed.